requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-six==1.0.4
filelock==3.0.12
fpdf==1.7.2
gunicorn==20.0.4
idna==2.10
img2pdf==0.4.0
importlib-metadata==3.4.0
importlib-resources==5.1.0
lxml==4.6.2
pandas==0.25.3
pikepdf==2.2.4
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.2
waitress==1.4.4
whitenoise==5.2.0
zipp==3.4.0
